-Add verification for CouponService.java methods -Add update in controller and in service

+++ src/main/java/ro/itschool/Booking/service/CouponService.java
@@ -2,11 +2,13 @@ package ro.itschool.Booking.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import ro.itschool.Booking.customException.IncorrectIdException;
 import ro.itschool.Booking.entity.Coupon;
 import ro.itschool.Booking.repository.CouponRepository;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -34,18 +36,49 @@ public class CouponService {
         return saveCoupon(coupon);
     }
 
-    public Coupon getCoupon(Long id) {
-        Optional<Coupon> getCouponById = couponRepository.findById(id);
-        Coupon coupon = getCouponById.get();
+    public Coupon update(Long id,String code, String validFrom_, String validTo_, double discount, boolean isActiv) throws Exception {
+        Coupon coupon = getCoupon(id);
+        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("uuuu/MM/dd");
+        LocalDate validFrom = LocalDate.parse(validFrom_, formatter);
+        LocalDate validTo = LocalDate.parse(validTo_, formatter);
+        coupon.setCode(code);
+        coupon.setValidFrom(validFrom);
+        coupon.setValidTo(validTo);
+        coupon.setDiscount(discount);
+        coupon.setActivCoupon(isActiv);
         return coupon;
     }
 
+    public Coupon getCoupon(Long id) throws IncorrectIdException {
+        if (id != null) {
+            Optional<Coupon> getCouponById = couponRepository.findById(id);
+            Coupon coupon = getCouponById.get();
+            return coupon;
+        } else {
+            throw new IncorrectIdException("This id: " + id + "was not found!");
+        }
+    }
+
     public Long deleteCoupon(Long id) {
-        couponRepository.deleteById(id);
+        try {
+            if (id != null) {
+                couponRepository.deleteById(id);
+                return id;
+            } else {
+                throw new IncorrectIdException("This id: " + id + " was not found");
+            }
+        } catch (IncorrectIdException e) {
+            e.printStackTrace();
+        }
         return id;
     }
 
     public List<Coupon> getAllCoupons() {
-        return couponRepository.findAll();
+        List<Coupon> allCoupons = couponRepository.findAll();
+        if (allCoupons.isEmpty()) {
+            return new ArrayList<>();
+        } else {
+            return allCoupons;
+        }
     }
 }
